Refactor into single threataggregator package
